requirements.txt
opencv-python
deepface
torch
torchvision
diffusers
Pillow
streamlit
